// src/lib/data/types.ts
export type Type = {
	name:
		| 'Normal'
		| 'Fighting'
		| 'Flying'
		| 'Poison'
		| 'Ground'
		| 'Rock'
		| 'Bug'
		| 'Ghost'
		| 'Steel'
		| 'Fire'
		| 'Water'
		| 'Grass'
		| 'Electric'
		| 'Psychic'
		| 'Ice'
		| 'Dragon'
		| 'Dark'
		| 'Fairy';
};

export const types: Type[] = [
	{
		name: 'Normal'
	},
	{
		name: 'Fighting'
	},
	{
		name: 'Flying'
	},
	{
		name: 'Poison'
	},
	{
		name: 'Ground'
	},
	{
		name: 'Rock'
	},
	{
		name: 'Bug'
	},
	{
		name: 'Ghost'
	},
	{
		name: 'Steel'
	},
	{
		name: 'Fire'
	},
	{
		name: 'Water'
	},
	{
		name: 'Grass'
	},
	{
		name: 'Electric'
	},
	{
		name: 'Psychic'
	},
	{
		name: 'Ice'
	},
	{
		name: 'Dragon'
	},
	{
		name: 'Dark'
	},
	{
		name: 'Fairy'
	}
];